deprecated old download code command
