+++ cli/src/commands/index.ts
@@ -1,4 +1,3 @@
-export * from "./download-code-diff-2";
 export * from "./download-code-command";
 export * from "./check-command";
 export * from "./storage-change-command";
